requirements.txt
numpy
pandas
scikit-learn
opencv-python
matplotlib
seaborn
pillow
#tensorflow
mlflow
dvc
tqdm
pyyaml
loguru
pytest
python-dotenv
torch
torchvision
mlflow
dvc[azure]
apache-airflow
docker
docker-compose

